Add files via upload
I have completed this project for Speech and Audio Recognition module during my postgraduate study. The overarching goal of this coursework is to synthesize some vowels (monophthongs) employing the source-filter model of speech production. The whole procedure is completed by MATLAB software using different in-built functions. To complete the implementation, Linear predictive coding (LPC) is used to for identifying the formant frequencies where cepstrum is used to predict fundamental frequency. Finally, to synthesize the vowel samples a periodic impulse train is generated.